clean up (part 4): renaming, inner classes to top-level types, API doc
